services/gateway/pyproject.toml
[tool.poetry]
name = "clestiq-shield-gateway"
version = "1.0.0"
description = "Clestiq Shield Gateway Service"
authors = ["Admin <vasubhut157@gmail.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic-settings = "^2.1.0"
pydantic = "^2.0.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.25"
greenlet = "^3.0.3"
structlog = "^24.1.0"
ddtrace = "^2.0.0"
httpx = "0.27.0"
datadog = "^0.49.0"
redis = "^5.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
